Rename color utils file
