+++ src/chat/chatbox.html
@@ -27,7 +27,7 @@
 <body>
 	<ul data-twitch-chat="{{ eleventy.serverless.path.channel }}"></ul>
 	<!-- <script src="/scripts/bttvIntegration.mjs" type="module"></script> -->
-	<script src="/scripts/colorContrast.mjs" type="module"></script>
+	<script src="/scripts/color.mjs" type="module"></script>
 	<script src="/scripts/handleChat.mjs" type="module"></script>
 	<script src="/scripts/utilities.mjs" type="module"></script>
 </body>
